requirements.txt
selenium==4.25.0            # for selenium
pytest==8.3.3               # for pytest
pytest-order==1.3.0         # for test order
pytest-html==4.1.1          # for basic html report (optional)
ddt==1.7.2                  # for data driven testing (optional)
pytest-xdist==2.5.0         # for running tests in parallel (optional)
urllib3==1.26.7             # urllib3
allure-pytest==2.13.5       # for allure reports